+++ Homework/java-web/domain/src/main/java/com/konstantion/review/dto/CreationReviewDto.java
@@ -2,5 +2,5 @@ package com.konstantion.review.dto;
 
 import java.util.UUID;
 
-public record CreationReviewDto(String message, Integer rating, UUID productUuid) {
+public record CreationReviewDto(String message, Double rating, UUID productUuid) {
 }
