+++ Homework/java-web/domain/src/main/java/com/konstantion/review/dto/ReviewDto.java
@@ -3,5 +3,5 @@ package com.konstantion.review.dto;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record ReviewDto(UUID uuid, String message, Integer rating, UUID userUuid, UUID productUuid, LocalDateTime createdAt) {
+public record ReviewDto(UUID uuid, String message, Double rating, UUID userUuid, UUID productUuid, LocalDateTime createdAt) {
 }
